+++ packages/zent-breadcrumb/src/index.js
@@ -1,3 +1,1 @@
-import Breadcrumb from './Breadcrumb';
-
-export default Breadcrumb;
+export default from './Breadcrumb';