+++ packages/zent-form/src/form_components/RadioGroupField.js
@@ -1,10 +1,14 @@
 import React from 'react';
 import Radio from 'zent-radio';
 import getControlGroup from '../getControlGroup';
+import unknownProps from '../unknownProps';
+import omit from 'zent-utils/lodash/omit';
 
 const RadioGroup = Radio.Group;
-const RadioGroupField = getControlGroup(props => {
-  return <RadioGroup {...props} />;
-});
+const RadioGroupWrap = (props) => {
+  const passableProps = omit(props, unknownProps);
+  return <RadioGroup className="zent-form__radio-group" {...passableProps} />;
+};
+const RadioGroupField = getControlGroup(RadioGroupWrap);
 
 export default RadioGroupField;