+++ packages/zent-form/src/unknownProps.js
@@ -1,0 +1,14 @@
+// 原生标签不能传非标准属性，否则会有warning
+// https://facebook.github.io/react/warnings/unknown-prop.html
+export default [
+  'isTouched',
+  'isPristine',
+  'isValid',
+  'error',
+  'errors',
+  'validationError',
+  'validationErrors',
+  'validations',
+  'normalize',
+  'format'
+];