+++ packages/zent-menu/src/index.js
@@ -1,0 +1,8 @@
+import Menu from './Menu';
+import MenuItem from './MenuItem';
+import SubMenu from './SubMenu';
+
+Menu.MenuItem = MenuItem;
+Menu.SubMenu = SubMenu;
+
+export default Menu;