+++ packages/zent-select/src/triggers/SimpleTrigger.js
@@ -7,9 +7,9 @@ import React, { Component, PropTypes } from 'react';
 class SimpleTrigger extends Component {
 
   render() {
-    let { prefixCls } = this.props;
+    let { prefixCls, onClick } = this.props;
 
-    return <div {...this.props} className={`${prefixCls}-simple`}>{this.props.text || this.props.placeholder}</div>;
+    return <div className={`${prefixCls}-simple`} onClick={onClick}>{this.props.text || this.props.placeholder}</div>;
   }
 }